requirements.txt
torch
torchvision
transformers
datasets
evaluate
jiwer
accelerate
scikit-learn
pillow